Docker/frontend changes
