--- frontends/client_frontend/src/interfaces/OrdersPage.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import OrderHistory from '../components/OrderHistory';
-
-const OrdersPage = ({ auth }) => (
-  <div>
-    <OrderHistory token={auth?.accessToken} />
-  </div>
-);
-
-export default OrdersPage;
