+++ frontends/client_frontend/src/pages/CartPage.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Cart from '../components/Cart';
 
-const CartPage = ({ auth, cartItems, removeFromCart, onBuy }) => (
+const CartPage = ({ cartItems, removeFromCart, onBuy }) => (
   <div>
     <Cart cartItems={cartItems} removeFromCart={removeFromCart} onBuy={onBuy}/>
   </div>
