+++ frontends/client_frontend/src/pages/HomePage.js
