+++ frontends/client_frontend/src/pages/ProductPage.js
