+++ frontends/client_frontend/src/pages/RegisterPage.js
