+++ frontends/client_frontend/src/style/ProductList.css
@@ -1,6 +1,7 @@
 .product-list-container {
     max-width: 1000px;
     margin: 0 auto;
+    margin-top: 50px;
     padding: 20px;
     background-color: #f9f9f9;
     border-radius: 10px;
